Squash a heap of resource clone warnings
Because we call the actions in this provider multiple times, it leads to
a lot of clone resource warnings for all of the directories and files
that are created within that execute.

Use the "resource pattern" to create resources if they do not exist, and
exit existing resources if they do exist. This is the equiv of the
chef-edit gem code.